cleanup: move cache util
