+++ lib/stats.js
@@ -1,8 +1,8 @@
 import pMap from 'p-map'
 
 import mongo from './util/mongo.js'
+import cache from './util/cache.js'
 import {getPeriods, getYesterdayRange, getLastWeekRange, getLastMonthRange, getLastYearRange, getAllTimeRange, formatDate} from './dates.js'
-import cache from './cache.js'
 
 async function computePresenceStats(range) {
   const result = await mongo.db.collection('users').aggregate([
