+++ lib/util/cache.js
